Realocação dos scripts util
