+++ pom.xml
@@ -71,7 +71,11 @@
             <version>2.11.0</version>
 		</dependency>
 
-
+		<dependency>
+    		<groupId>javax.inject</groupId>
+   			 <artifactId>javax.inject</artifactId>
+    <version>1</version>
+</dependency>
 	</dependencies>
 
 	<build>
