+++ src/main/java/com/next2023/cloud_service/repositories/ArquivosRepoisitory.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import com.next2023.cloud_service.entities.Arquivos;
 
-@Repository
+//@Repository
 public interface ArquivosRepoisitory extends JpaRepository <Arquivos, Long> {
     
 }
