+++ src/main/java/controllers/UsuarioController.java
@@ -6,6 +6,8 @@ package controllers;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.inject.Inject;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
@@ -19,6 +21,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.next2023.cloud_service.entities.Usuario;
+import com.next2023.cloud_service.services.AwsConfigService;
+import com.next2023.cloud_service.util.AwsConfig;
 
 import jakarta.validation.Valid;
 
@@ -29,9 +33,13 @@ import jakarta.validation.Valid;
 
 public class UsuarioController {
 
+@Inject
+    public AwsConfigService awsConfigService; // usado para teste. remover
+
     @GetMapping
     public ResponseEntity<List<Usuario>> listAll(){
         List<Usuario> listUsuarios = new ArrayList<Usuario>();
+        awsConfigService.listBuckets();
         return new ResponseEntity<List<Usuario>>(listUsuarios, HttpStatus.OK);
     }
