+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+awsS3.accessKey = AKIAW5KS6GMSREW3FFB7
+awsS3.secretKey = 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
